requirements.txt.txt
kornia@git+https://github.com/kornia/kornia@77589a58be6c603b7afd755d261783bd0c152a97
matplotlib
Shapely==1.8.0
easydict
pytorch-lightning==1.5.8
loguru 
scipy
h5py
mmcv-full==1.7.0
mmdet==2.28.2
wandb
segmentation-models-pytorch@git+https://github.com/ragavsachdeva/segmentation_models.pytorch.git@0092ee4d6f851d89a4a401bb2dfa6187660b8dd3
imageio==2.13.5
timm==0.6.12 
jsonargparse 
imageio 
einops 
etils 
